Move compacted_db_impl.[c|h] to db/db_impl
