+++ db/db_impl/compacted_db_impl.cc
@@ -4,7 +4,7 @@
 //  (found in the LICENSE.Apache file in the root directory).
 
 #ifndef ROCKSDB_LITE
-#include "db/compacted_db_impl.h"
+#include "db/db_impl/compacted_db_impl.h"
 
 #include "db/db_impl/db_impl.h"
 #include "db/version_set.h"
